+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Possible log types:
 
 - [changed] Rename `onDisconnected` in task interface to `onPeerDisconnected`
 - [fixed] Only call `onPeerDisconnected` if defined by the task
+- [fixed] Accept server messages during/after peer handshake
 
 ### v0.11.0 (2018-03-13)